Define class DSL from within its own module subclass
